# api-gateway/ibanking-swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/user/register": {
      "post": {
        "description": "Create a new user account",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "User failed to be created"
          }
        },
        "summary": "Register user",
        "tags": [
          "User"
        ]
      }
    },
    "/user/login": {
      "post": {
        "description": "User login with email and password",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Email or password is incorrect",
                    "error": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          }
        },
        "summary": "User login",
        "tags": [
          "User"
        ]
      }
    },
    "/user/users": {
      "get": {
        "description": "Return all users in the system",
        "operationId": "AuthController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Users fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch users"
          }
        },
        "summary": "Get all users",
        "tags": [
          "User"
        ]
      }
    },
    "/user/{id}/deduct-balance": {
      "post": {
        "description": "Subtract a certain amount from the user balance",
        "operationId": "AuthController_deductBalance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeductBalanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Balance deducted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceDeductResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Not enough balance"
          }
        },
        "summary": "Deduct user balance",
        "tags": [
          "User"
        ]
      }
    },
    "/user/{id}/refund-balance": {
      "post": {
        "description": "Increase user balance by a certain amount (refund case)",
        "operationId": "AuthController_refundBalance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeductBalanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Balance refunded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Balance refunded",
                    "newBalance": 1500000
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Refund user balance",
        "tags": [
          "User"
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "description": "Return user details by ID",
        "operationId": "AuthController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserIDResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Get user by ID",
        "tags": [
          "User"
        ]
      }
    },
...
    "/otp/generate": {
      "post": {
        "description": "Create a new OTP and send it via email",
        "operationId": "OtpController_generate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OTP generated and sent to email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateOtpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Failed to generate or send OTP"
          }
        },
        "summary": "Generate OTP",
        "tags": [
          "OTP"
        ]
      }
    },
    "/otp/verify": {
      "post": {
        "description": "Verify the OTP entered by the user",
        "operationId": "OtpController_verify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OTP verification result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyOtpResponseDto"
                }
              }
            }
          }
        },
        "summary": "Verify OTP",
        "tags": [
          "OTP"
        ]
      }
    }
  },
  "info": {
    "title": "IBanking Documentation",
    "description": "Combined API Docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "student@gmail.com"
          },
          "name": {
            "type": "string",
            "example": "Nguyen Van A"
          },
          "password": {
            "type": "string",
            "example": "secret123"
          },
          "phone": {
            "type": "string",
            "example": "0123456789",
            "default": ""
          },
          "balance": {
            "type": "number",
            "example": 1000.5,
            "default": 0
          }
        },
        "required": [
          "email",
          "password",
          "phone",
          "balance"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "email": {
            "type": "string",
            "example": "student@example.com"
          },
          "name": {
            "type": "string",
            "example": "Nguyen Van A"
          },
          "phone": {
            "type": "string",
            "example": "0123456789"
          },
          "balance": {
            "type": "number",
            "example": 1000.5
          },
          "createdAt": {
            "type": "string",
            "example": "2025-09-19T13:55:55.006Z"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "phone",
          "balance",
          "createdAt"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "student@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "secret123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Login successful"
          },
          "student": {
            "$ref": "#/components/schemas/StudentResponseDto"
          }
        },
        "required": [
          "message",
          "student"
        ]
      },
      "DeductBalanceDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 100
          }
        },
        "required": [
          "amount"
        ]
      },
      "BalanceDeductResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Balance deducted"
          },
          "newBalance": {
            "type": "number",
            "example": 1000.5
          }
        },
        "required": [
          "message",
          "newBalance"
        ]
      },
      "UserIDResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "email": {
            "type": "string",
            "example": "student@example.com"
          },
          "name": {
            "type": "string",
            "example": "Nguyen Van A"
          },
          "phone": {
            "type": "string",
            "example": "0123456789"
          },
          "balance": {
            "type": "number",
            "example": 1000.5
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "phone",
          "balance"
        ]
      },
...
      "StudentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "sID": {
            "type": "string",
            "example": "S12345"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "student@example.com"
          },
          "phoneNumber": {
            "type": "string",
            "example": "0123456789"
          },
          "address": {
            "type": "string",
            "example": "New York"
          },
          "dateOfBirth": {
            "format": "date-time",
            "type": "string",
            "example": "2002-10-10T00:00:00.000Z",
            "description": "Date of birth (ISO 8601)"
          }
        },
        "required": [
          "id",
          "sID",
          "name"
        ]
      },
...
      "GenerateOtpDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User ID in the system",
            "example": "user_12345"
          },
          "email": {
            "type": "string",
            "description": "User email to send OTP",
            "example": "student@example.com"
          },
          "checkoutId": {
            "type": "string",
            "description": "Checkout ID (optional)",
            "example": "checkout-456"
          }
        },
        "required": [
          "userId",
          "email"
        ]
      },
      "GenerateOtpResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "76f3f4ca-73b1-4113-9e3f-76dc33d26f7a"
          },
          "code": {
            "type": "string",
            "example": "981410"
          },
          "userId": {
            "type": "string",
            "example": "user_12345"
          },
          "transactionId": {
            "type": "string",
            "example": "txn-456"
          },
          "expiresAt": {
            "type": "string",
            "example": "2025-09-19T14:15:56.312Z"
          },
          "used": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "example": "2025-09-19T14:10:56.315Z"
          }
        },
        "required": [
          "id",
          "code",
          "userId",
          "expiresAt",
          "used",
          "createdAt"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User ID in the system",
            "example": "user_12345"
          },
          "code": {
            "type": "string",
            "description": "OTP code entered by the user",
            "example": "123456"
          },
          "checkoutId": {
            "type": "string",
            "description": "Checkout ID associated with the OTP",
            "example": "checkout_12345"
          }
        },
        "required": [
          "userId",
          "code",
          "checkoutId"
        ]
      },
      "VerifyOtpResponseDto": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "valid"
        ]
      }
    }
  }
}